60FPS, Saving and Loading Working Except NPCs
Numerous bug fixes and performance improvements.  Can now run 60FPS and saving and loading no longer crashes.  Still bugs with other items however such as keypad UI and NPC loading.

+++ Assets/Shaders/UnityDeferredLibrary.cginc
@@ -154,7 +154,8 @@ void UnityGGXDeferredCalculateLightParams (
     // read depth and reconstruct world position
     float depth = SAMPLE_DEPTH_TEXTURE(_CameraDepthTexture, uv);
     depth = Linear01Depth (depth);
-    float4 vpos = float4(i.ray * depth,1);
+    //float4 vpos = float4(i.ray * depth,1);
+    half4 vpos = float4(i.ray * depth,1);
     float3 wpos = mul (unity_CameraToWorld, vpos).xyz;
 
     float fadeDist = UnityComputeShadowFadeDistance(wpos, vpos.z);
